Updated dnie-tool.1.xml for consistency
Experienced a problem with dnie-tool where I would receive a warning with the content of /etc/bash_completion.d/dnie-tool.

The cause of the error was a missing case label which in turn was caused by the formatting of the dnie-tool.1.xml.

Options were formatted like <term><option--xarg, -x</option></term> which were not handled by the sed regular expression in the makefiles.

Modified the dnie-tool.1.xml file to be consistent with the other doc files and to generate the dnie-tool file correctly.